+++ lib/file-source/src/file_watcher.rs
@@ -27,7 +27,7 @@ impl FileWatcher {
     ///
     /// The input path will be used by `FileWatcher` to prime its state
     /// machine. A `FileWatcher` tracks _only one_ file. This function returns
-    /// None if the path does not exist or is not readable by cernan.
+    /// None if the path does not exist or is not readable by the current process.
     pub fn new(
         path: PathBuf,
         file_position: FilePosition,