+++ scripts/generate/context/config_example.rb
@@ -26,9 +26,9 @@ class Context
     def tags(option)
       tags = []
 
-      if !option.default.nil?
+      if option.examples.first == option.default
         tags << "default"
-      elsif option.optional?
+      elsif option.default.nil? && option.optional?
         tags << "no default"
       end